+++ modules/ai/ai.py
@@ -2,7 +2,7 @@ import random
 
 from pygame import Rect, draw
 
-from square import CATCH_COLOR, FLEE_COLOR, Square
+from modules.objects.square import CATCH_COLOR, FLEE_COLOR, Square
 
 
 class Ai(Square):
