+++ modules/settings.py
@@ -1,0 +1,3 @@
+FPS_LIMIT = 250
+RUNNERS = 8
+CATCHERS = 8
