+++ modules/states/game.py
@@ -6,9 +6,9 @@ os.environ["PYGAME_HIDE_SUPPORT_PROMPT"] = "1"
 
 import pygame
 
-from modules.menu import MainMenuState, PauseMenuState
-from modules.state import GameState
-from play import RunningGameState
+from modules.states.menu import MainMenuState, PauseMenuState
+from modules.states.play import RunningGameState
+from modules.states.state import GameState
 
 # Stałe wartości
 SCREEN_WIDTH = 1600  # Szerokość ekranu
